セルにスタイルを適用する処理と色を暗くする処理を切り出し
