+++ app/javascript/controllers/slider_controller.js
@@ -1,7 +1,11 @@
 import { Controller } from "@hotwired/stimulus"
+import { CellStyles } from "./cell_styles"
 
 export default class extends Controller {
   static targets = ["value", "colorSample"]
+
+  applyStyles = CellStyles.applyStyles.bind(this)
+  darkenColorValue = CellStyles.darkenColorValue.bind(this)
 
   connect() {
     this.changeSample(this.valueTarget.textContent)
@@ -40,24 +44,5 @@ export default class extends Controller {
       case 2:
         return 0.6
     }
-  }
-
-  applyStyles(cell, color, opacity = 1) {
-    cell.style.backgroundColor = color
-    cell.style.opacity = opacity
-  }
-
-  darkenColorValue(hex) {
-    const darkenValue = 65
-
-    const rgb = [
-      parseInt(hex.slice(1, 3), 16),
-      parseInt(hex.slice(3, 5), 16),
-      parseInt(hex.slice(5, 7), 16)
-    ]
-
-    const darkenedRgb = rgb.map(value => Math.max(0, value - darkenValue))
-
-    return `rgb(${darkenedRgb.join(", ")})`
   }
 }
